bug-triage-system/requirements.txt
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
schedule==1.2.0
